+++ bgstally/ui.py
@@ -9,7 +9,6 @@ from tkinter.messagebox import askyesno
 from typing import List, Optional
 
 import myNotebook as nb
-import semantic_version
 from ttkHyperlinkLabel import HyperlinkLabel
 
 from bgstally.activity import Activity
@@ -46,6 +45,7 @@ class UI:
         self.image_button_carrier = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "button_carrier.png"))
 
         self.indicate_activity:bool = False
+        self.report_system_address:str = None
 
         # Single-instance windows
         self.window_cmdrs:WindowCMDRs = WindowCMDRs(self.bgstally)
@@ -175,6 +175,14 @@ class UI:
         return frame
 
 
+    def show_system_report(self, system_address:int):
+        """
+        Show the system report overlay
+        """
+        self.indicate_activity = True
+        self.report_system_address = str(system_address)
+
+
     def _worker(self) -> None:
         """
         Handle thread work for overlay
@@ -213,6 +221,12 @@ class UI:
                     percent:float = round(progress * 100, 2)
 
                     self.bgstally.overlay.display_progress_bar("tw", f"TW War Progress in {current_system.get('System', 'Unknown')}: {percent}%", progress)
+
+            if self.report_system_address is not None and current_activity is not None:
+                report_system:dict = current_activity.get_system_by_address(self.report_system_address)
+                if report_system is not None:
+                    self.bgstally.overlay.display_message("system_info", current_activity.generate_text(DiscordActivity.BOTH, False, report_system['System']), fit_to_text=True, has_title=True)
+                self.report_system_address = None
 
             sleep(TIME_WORKER_PERIOD_S)