+++ bgstally/utils.py
@@ -9,3 +9,15 @@ def get_by_path(dic:dict, keys:list, default:any = None):
         return default
 
     return dic
+
+
+def human_format(num):
+    """
+    Format a BGS value into shortened human-readable text
+    """
+    num = float('{:.3g}'.format(num))
+    magnitude = 0
+    while abs(num) >= 1000:
+        magnitude += 1
+        num /= 1000.0
+    return '{}{}'.format('{:f}'.format(num).rstrip('0').rstrip('.'), ['', 'K', 'M', 'B', 'T'][magnitude])